+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   - Create new page for company subscriptions and add subscription management functionality
 - Quick links
   - update props for quick links and updated shared components package
+- Company Wallet
+  - Card style updates for Active, Inactive and Unknown status
 
 ### Bugfix